Creating ability to mark days as undoable
Rescheduling after marking day as undoable done through simplistic algorithm.
Algorithm searches for a nearby day, in either direction, as long as it is
in the future, with a progressively expanding scope.

It also ensures that Hero being swapped with has not marked the given day as
undoable by collecting hero_ids from the recorded undoable days for that day.